backend/requirements.txt
flask==2.0.1
waitress>=2.0.0
weaviate-client==4.5.0
sentence-transformers==2.2.2
requests>=2.0.0
python-dotenv==0.19.1
numpy==1.24.3
scikit-learn>=1.0.0
transformers>=4.0.0
llama-cpp-python==0.1.65
pyyaml==6.0 